[BACKLOG-26905] - jms consumer might lose messages on safe stopping
need to wait for receive loop to finish before completing publishProcessor

rx Observable created multiple threads so that emitter.onNext was asyncronous before hitting the publisher.  that allowed time for the publisher to close before it got the last message sent.  removed in favor of a single thread for receive loop